requirements.txt
anthropic>=0.39.0
openai>=1.50.0
google-genai>=0.3.0
notion-client>=2.2.1
supabase>=2.10.0
praw>=7.8.0
feedparser>=6.0.11
google-api-python-client>=2.150.0
google-auth>=2.36.0
google-auth-oauthlib>=1.2.1
google-auth-httplib2>=0.2.0
python-dotenv>=1.0.1
requests>=2.32.0
Pillow>=11.0.0
